Synchronize install_compiler.sh with the trunk
git-svn-id: http://svn.open64.net/svnroot/open64@2994 ead328b1-e903-0410-a19d-eacee4b031f6

+++ branches/open64-4.2.3/install_compiler.sh
@@ -204,12 +204,29 @@ INSTALL_GCC () {
 
     # Make links to gcc runtime libraries
     cd ${ROOT}
-    for i in open64-gcc-4.2.0/lib64/lib*.so*; do
-        (cd $PHASEPATH; ln -sf ../../../../$i `basename $i`)
-    done
-    for i in open64-gcc-4.2.0/lib/lib*.so*; do
-        (cd $PHASEPATH/32; ln -sf ../../../../../$i `basename $i`)
-    done
+    if [ "$TARG_HOST" = "x8664" ]
+    then
+	for i in open64-gcc-4.2.0/lib64/lib*.so*; do
+	    if [ -e "$i" ]
+	    then
+		(cd $PHASEPATH; ln -sf ../../../../$i `basename $i`)
+	    fi
+	done
+	for i in open64-gcc-4.2.0/lib/lib*.so*; do
+	    if [ -e "$i" ]
+	    then
+		(cd $PHASEPATH/32; ln -sf ../../../../../$i `basename $i`)
+	    fi
+	done
+    elif [ "$TARG_HOST" = "ia64" ]
+    then
+	for i in open64-gcc-4.2.0/lib/lib*.so*; do
+	    if [ -e "$i" ]
+	    then
+		(cd $PHASEPATH; ln -sf ../../../../$i `basename $i`)
+	    fi
+	done
+    fi
 
     popd
